+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Render markdown links in text and list input fields
+
 ## [0.2.0] - 2026-08-09
 
 ### Added